+++ baybe/surrogates/gaussian_process/core.py
@@ -5,8 +5,10 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, ClassVar
 
 from attrs import define, field
+from attrs.validators import instance_of
 
-from baybe.searchspace import SearchSpace
+from baybe.parameters.base import Parameter
+from baybe.searchspace.core import SearchSpace
 from baybe.surrogates.base import Surrogate
 from baybe.surrogates.gaussian_process.kernel_factory import (
     KernelFactory,
@@ -23,12 +25,68 @@ from baybe.surrogates.gaussian_process.presets.default import (
 
 if TYPE_CHECKING:
     from botorch.models.model import Model
+    from botorch.models.transforms.input import InputTransform
+    from botorch.models.transforms.outcome import OutcomeTransform
+    from botorch.posteriors import Posterior
     from torch import Tensor
+
+
+@define
+class _ModelContext:
+    """Model context for :class:`GaussianProcessSurrogate`."""
+
+    searchspace: SearchSpace = field(validator=instance_of(SearchSpace))
+    """The search space the model is trained on."""
+
+    @property
+    def task_idx(self) -> int | None:
+        """The computational column index of the task parameter, if available."""
+        return self.searchspace.task_idx
+
+    @property
+    def is_multitask(self) -> bool:
+        """Indicates if model is to be operated in a multi-task context."""
+        return self.n_task_dimensions > 0
+
+    @property
+    def n_task_dimensions(self) -> int:
+        """The number of task dimensions."""
+        # TODO: Generalize to multiple task parameters
+        return 1 if self.task_idx is not None else 0
+
+    @property
+    def n_tasks(self) -> int:
+        """The number of tasks."""
+        return self.searchspace.n_tasks
+
+    @property
+    def parameter_bounds(self) -> Tensor:
+        """Get the search space parameter bounds in BoTorch Format."""
+        import torch
+
+        return torch.from_numpy(self.searchspace.comp_rep_bounds.values)
+
+    def get_numerical_indices(self, n_inputs: int) -> tuple[int, ...]:
+        """Get the indices of the regular numerical model inputs."""
+        return tuple(i for i in range(n_inputs) if i != self.task_idx)
 
 
 @define
 class GaussianProcessSurrogate(Surrogate):
     """A Gaussian process surrogate model."""
+
+    # Note [Scaling Workaround]
+    # -------------------------
+    # For GPs, we deactivate the base class scaling and instead let the botorch
+    # model internally handle input/output scaling. The reasons is that we need to
+    # make `to_botorch` expose the actual botorch GP object, instead of going
+    # via the `AdapterModel`, because certain acquisition functions (like qNIPV)
+    # require the capability to `fantasize`, which the `AdapterModel` does not support.
+    # The base class scaling thus needs to be disabled since otherwise the botorch GP
+    # object would be trained on pre-scaled input/output data. This would cause a
+    # problem since the resulting `posterior` method of that object is exposed
+    # to `optimize_acqf_*`, which is configured to be called on the original scale.
+    # Moving the scaling operation into the botorch GP object avoids this conflict.
 
     # Class variables
     joint_posterior: ClassVar[bool] = True
@@ -55,7 +113,7 @@ class GaussianProcessSurrogate(Surrogate):
     _model = field(init=False, default=None, eq=False)
     """The actual model."""
 
-    @classmethod
+    @staticmethod
     def from_preset(preset: GaussianProcessPreset) -> GaussianProcessSurrogate:
         """Create a Gaussian process surrogate from one of the defined presets."""
         return make_gp_from_preset(preset)
@@ -65,34 +123,46 @@ class GaussianProcessSurrogate(Surrogate):
 
         return self._model
 
-    def _posterior(self, candidates: Tensor) -> tuple[Tensor, Tensor]:
+    @staticmethod
+    def _make_parameter_scaler_factory(
+        parameter: Parameter,
+    ) -> type[InputTransform] | None:
         # See base class.
-        posterior = self._model.posterior(candidates)
-        return posterior.mvn.mean, posterior.mvn.covariance_matrix
 
-    def _fit(self, searchspace: SearchSpace, train_x: Tensor, train_y: Tensor) -> None:
+        # For GPs, we let botorch handle the scaling. See [Scaling Workaround] above.
+        return None
+
+    @staticmethod
+    def _make_target_scaler_factory() -> type[OutcomeTransform] | None:
+        # See base class.
+
+        # For GPs, we let botorch handle the scaling. See [Scaling Workaround] above.
+        return None
+
+    def _posterior(self, candidates_comp_scaled: Tensor, /) -> Posterior:
+        # See base class.
+        return self._model.posterior(candidates_comp_scaled)
+
+    def _fit(self, train_x: Tensor, train_y: Tensor) -> None:
         # See base class.
 
         import botorch
         import gpytorch
         import torch
 
-        # identify the indexes of the task and numeric dimensions
-        # TODO: generalize to multiple task parameters
-        task_idx = searchspace.task_idx
-        n_task_params = 1 if task_idx is not None else 0
-        numeric_idxs = [i for i in range(train_x.shape[1]) if i != task_idx]
+        # FIXME[typing]: It seems there is currently no better way to inform the type
+        #   checker that the attribute is available at the time of the function call
+        assert self._searchspace is not None
 
-        # get the input bounds from the search space in BoTorch Format
-        bounds = torch.from_numpy(searchspace.param_bounds_comp)
-        # TODO: use target value bounds when explicitly provided
+        context = _ModelContext(self._searchspace)
 
-        # define the input and outcome transforms
-        # TODO [Scaling]: scaling should be handled by search space object
+        numerical_idxs = context.get_numerical_indices(train_x.shape[-1])
+
+        # For GPs, we let botorch handle the scaling. See [Scaling Workaround] above.
         input_transform = botorch.models.transforms.Normalize(
-            train_x.shape[1], bounds=bounds, indices=numeric_idxs
+            train_x.shape[-1], bounds=context.parameter_bounds, indices=numerical_idxs
         )
-        outcome_transform = botorch.models.transforms.Standardize(train_y.shape[1])
+        outcome_transform = botorch.models.transforms.Standardize(train_y.shape[-1])
 
         # extract the batch shape of the training data
         batch_shape = train_x.shape[:-2]
@@ -102,26 +172,26 @@ class GaussianProcessSurrogate(Surrogate):
 
         # define the covariance module for the numeric dimensions
         base_covar_module = self.kernel_factory(
-            searchspace, train_x, train_y
+            context.searchspace, train_x, train_y
         ).to_gpytorch(
-            ard_num_dims=train_x.shape[-1] - n_task_params,
-            active_dims=numeric_idxs,
+            ard_num_dims=train_x.shape[-1] - context.n_task_dimensions,
+            active_dims=numerical_idxs,
             batch_shape=batch_shape,
         )
 
         # create GP covariance
-        if task_idx is None:
+        if not context.is_multitask:
             covar_module = base_covar_module
         else:
             task_covar_module = gpytorch.kernels.IndexKernel(
-                num_tasks=searchspace.n_tasks,
-                active_dims=task_idx,
-                rank=searchspace.n_tasks,  # TODO: make controllable
+                num_tasks=context.n_tasks,
+                active_dims=context.task_idx,
+                rank=context.n_tasks,  # TODO: make controllable
             )
             covar_module = base_covar_module * task_covar_module
 
         # create GP likelihood
-        noise_prior = _default_noise_factory(searchspace, train_x, train_y)
+        noise_prior = _default_noise_factory(context.searchspace, train_x, train_y)
         likelihood = gpytorch.likelihoods.GaussianLikelihood(
             noise_prior=noise_prior[0].to_gpytorch(), batch_shape=batch_shape
         )
@@ -142,7 +212,7 @@ class GaussianProcessSurrogate(Surrogate):
         # TODO: This is a simple temporary workaround to avoid model overfitting
         #   via early stopping in the presence of task parameters, which currently
         #   have no prior configured.
-        if n_task_params > 0:
+        if context.n_task_dimensions > 0:
             botorch.optim.fit.fit_gpytorch_mll_torch(mll, step_limit=200)
         else:
             botorch.fit.fit_gpytorch_mll(mll)