+++ baybe/surrogates/utils.py
@@ -2,58 +2,23 @@
 
 from __future__ import annotations
 
+import typing
 from collections.abc import Callable
 from functools import wraps
 from typing import TYPE_CHECKING
 
-from baybe.scaler import DefaultScaler
-from baybe.searchspace import SearchSpace
-
 if TYPE_CHECKING:
+    from botorch.posteriors import Posterior
     from torch import Tensor
 
     from baybe.surrogates.base import Surrogate
+    from baybe.surrogates.naive import MeanPredictionSurrogate
 
 
-def _prepare_inputs(x: Tensor) -> Tensor:
-    """Validate and prepare the model input.
-
-    Args:
-        x: The "raw" model input.
-
-    Returns:
-        The prepared input.
-
-    Raises:
-        ValueError: If the model input is empty.
-    """
-    from baybe.utils.torch import DTypeFloatTorch
-
-    if len(x) == 0:
-        raise ValueError("The model input must be non-empty.")
-    return x.to(DTypeFloatTorch)
-
-
-def _prepare_targets(y: Tensor) -> Tensor:
-    """Validate and prepare the model targets.
-
-    Args:
-        y: The "raw" model targets.
-
-    Returns:
-        The prepared targets.
-
-    Raises:
-        NotImplementedError: If there is more than one target.
-    """
-    from baybe.utils.torch import DTypeFloatTorch
-
-    if y.shape[1] != 1:
-        raise NotImplementedError(
-            "The model currently supports only one target or multiple targets in "
-            "DESIRABILITY mode."
-        )
-    return y.to(DTypeFloatTorch)
+_constant_target_model_store: dict[int, MeanPredictionSurrogate] = {}
+"""Dictionary for storing constant target fallback models. Keys are the IDs of the
+surrogate models that temporarily have a fallback attached because they were
+trained on constant training targets. Values are the corresponding fallback models."""
 
 
 def catch_constant_targets(cls: type[Surrogate], std_threshold: float = 1e-6):
@@ -65,47 +30,33 @@ def catch_constant_targets(cls: type[Surrogate], std_threshold: float = 1e-6):
 
     The modified class handles the above cases separately from "regular operation"
     by resorting to a :class:`baybe.surrogates.naive.MeanPredictionSurrogate`,
-    which is stored as an additional temporary attribute in its objects.
+    which is stored outside the model in a dictionary maintained by this decorator.
 
     Args:
         cls: The :class:`baybe.surrogates.base.Surrogate` to be augmented.
         std_threshold: The standard deviation threshold below which operation is
             switched to the alternative model.
 
-    Raises:
-        ValueError: If the class already contains an attribute with the same name
-            as the temporary attribute to be added.
-
     Returns:
         The modified class.
     """
-    # Name of the attribute added to store the alternative model
-    injected_model_attr_name = "_constant_target_model"
-
-    if injected_model_attr_name in (attr.name for attr in cls.__attrs_attrs__):
-        raise ValueError(
-            f"Cannot apply '{catch_constant_targets.__name__}' because "
-            f"'{cls.__name__}' already has an attribute '{injected_model_attr_name}' "
-            f"defined."
-        )
-
     from baybe.surrogates.naive import MeanPredictionSurrogate
 
     # References to original methods
     _fit_original = cls._fit
     _posterior_original = cls._posterior
 
-    def _posterior_new(self, candidates: Tensor) -> tuple[Tensor, Tensor]:
+    def _posterior_new(self, candidates_comp_scaled: Tensor, /) -> Posterior:
+        """Use fallback model if it exists, otherwise call original posterior."""
         # Alternative model fallback
-        if hasattr(self, injected_model_attr_name):
-            return getattr(self, injected_model_attr_name)._posterior(candidates)
+        if constant_target_model := _constant_target_model_store.get(id(self), None):
+            return constant_target_model._posterior(candidates_comp_scaled)
 
         # Regular operation
-        return _posterior_original(self, candidates)
+        return _posterior_original(self, candidates_comp_scaled)
 
-    def _fit_new(
-        self, searchspace: SearchSpace, train_x: Tensor, train_y: Tensor
-    ) -> None:
+    def _fit_new(self, train_x: Tensor, train_y: Tensor) -> None:
+        """Original fit but with fallback model creation for constant targets."""
         if not (train_y.ndim == 2 and train_y.shape[-1] == 1):
             raise NotImplementedError(
                 "The current logic is only implemented for single-target surrogates."
@@ -114,86 +65,24 @@ def catch_constant_targets(cls: type[Surrogate], std_threshold: float = 1e-6):
         # Alternative model fallback
         if train_y.numel() == 1 or train_y.std() < std_threshold:
             model = MeanPredictionSurrogate()
-            model._fit(searchspace, train_x, train_y)
-            try:
-                setattr(self, injected_model_attr_name, model)
-            except AttributeError as ex:
-                raise TypeError(
-                    f"'{catch_constant_targets.__name__}' is only applicable to "
-                    f"non-slotted classes but '{cls.__name__}' is a slotted class."
-                ) from ex
+            model._fit(train_x, train_y)
+            _constant_target_model_store[id(self)] = model
 
         # Regular operation
         else:
-            if hasattr(self, injected_model_attr_name):
-                delattr(self, injected_model_attr_name)
-            _fit_original(self, searchspace, train_x, train_y)
+            _constant_target_model_store.pop(id(self), None)
+            _fit_original(self, train_x, train_y)
 
     # Replace the methods
-    cls._posterior = _posterior_new
-    cls._fit = _fit_new
+    cls._posterior = _posterior_new  # type: ignore
+    cls._fit = _fit_new  # type: ignore
 
     return cls
 
 
-def autoscale(cls: type[Surrogate]):
-    """Make a ``Surrogate`` class automatically scale the domain it operates on.
-
-    More specifically, the modified class transforms its inputs before processing them
-    and untransforms the results before returning them. The fitted scaler used for these
-    transformations is stored in the class' objects as an additional temporary
-    attribute.
-
-    Args:
-        cls: The :class:`baybe.surrogates.base.Surrogate` to be augmented.
-
-    Raises:
-        ValueError: If the class already contains an attribute with the same name
-            as the temporary attribute to be added.
-
-    Returns:
-        The modified class.
-    """
-    # Name of the attribute added to store the scaler
-    injected_scaler_attr_name = "_autoscaler"
-
-    if injected_scaler_attr_name in (attr.name for attr in cls.__attrs_attrs__):
-        raise ValueError(
-            f"Cannot apply '{autoscale.__name__}' because "
-            f"'{cls.__name__}' already has an attribute '{injected_scaler_attr_name}' "
-            f"defined."
-        )
-
-    # References to original methods
-    _fit_original = cls._fit
-    _posterior_original = cls._posterior
-
-    def _posterior_new(self, candidates: Tensor) -> tuple[Tensor, Tensor]:
-        scaled = getattr(self, injected_scaler_attr_name).transform(candidates)
-        mean, covar = _posterior_original(self, scaled)
-        return getattr(self, injected_scaler_attr_name).untransform(mean, covar)
-
-    def _fit_new(
-        self, searchspace: SearchSpace, train_x: Tensor, train_y: Tensor
-    ) -> None:
-        scaler = DefaultScaler(searchspace.discrete.comp_rep)
-        scaled_x, scaled_y = scaler.fit_transform(train_x, train_y)
-        try:
-            setattr(self, injected_scaler_attr_name, scaler)
-        except AttributeError as ex:
-            raise TypeError(
-                f"'{autoscale.__name__}' is only applicable to "
-                f"non-slotted classes but '{cls.__name__}' is a slotted class."
-            ) from ex
-        _fit_original(self, searchspace, scaled_x, scaled_y)
-
-    # Replace the methods
-    cls._posterior = _posterior_new
-    cls._fit = _fit_new
-
-    return cls
-
-
+# FIXME[typing]: Typing should be reactivated once the `joint_posterior` attribute
+#   has been refactored/removed
+@typing.no_type_check
 def batchify(
     posterior: Callable[[Surrogate, Tensor], tuple[Tensor, Tensor]],
 ) -> Callable[[Surrogate, Tensor], tuple[Tensor, Tensor]]: